TRQ-3142 Increase maximum time for $prologalarm
$prologalarm was ignored on the prologue for a job. Also when
the epilogue was run the $prologalarm value was
ignored if it was more the 300. Now the $prologalarm
value is always honored regardless of how large it is for both
prologue and epilogue scripts. The default timeout is still
300 seconds.

Conflicts:
	src/resmom/prolog.c
	src/server/test/job_func/test_job_func.c

+++ src/resmom/test/prolog/scaffolding.c
@@ -28,7 +28,7 @@ char log_buffer[LOG_BUF_SIZE];
 int MOMPrologTimeoutCount;
 uid_t pbsuser;
 gid_t pbsgroup;
-unsigned int     pe_alarm_time = PBS_PROLOG_TIME;
+long     pe_alarm_time = PBS_PROLOG_TIME;
 char *path_epilogp;
 long TJobStartTimeout = 300;
